app/src/main/java/com/example/user/movierankings/Movie.java
package com.example.user.movierankings;

/**
 * Created by user on 07/08/2017.
 */

public class Movie {

    private String title;
    private String genre;
    private int ranking;

    public Movie(String title, String genre, int ranking){
        this.title = title;
        this.genre = genre;
        this.ranking = ranking;
    }

    public String getTitle(){
        return title;
    }

    public void setTitle(String title){
        this.title = title;
    }

    public String getMovieGenre(){
        return genre;
    }

    public void setGenre(String genre){
        this.genre = genre;
    }

    public int getRanking(){
        return ranking;
    }

    public void setRanking(int ranking){
        this.ranking = ranking;
    }

    @Override public String toString(){
        return "Title: " + this.title + ", Genre: " + this.genre + ", Ranking: " + this.ranking;
    }
}